requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
anthropic>=0.40.0
python-dotenv==1.0.0
pydantic>=2.5.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
msal==1.26.0
requests==2.31.0
APScheduler==3.10.4
supabase>=2.0.0
